logging-tutorial/src/create_log_bucket.py
import os
import boto3
from dotenv import load_dotenv

def main():

    load_dotenv()

    bucket_name = os.getenv("AWS_BUCKET_NAME")

    s3 = boto3.client(
        "s3",
        aws_access_key_id=os.getenv("AWS_ACCESS_KEY_ID"),
        aws_secret_access_key=os.getenv("AWS_SECRET_ACCESS_KEY"),
        region_name=os.getenv("AWS_REGION"),
    )

    s3.create_bucket(
        Bucket=bucket_name,
        CreateBucketConfiguration={"LocationConstraint": os.getenv("AWS_REGION")},
    )

if __name__ == "__main__":
    main()
