requirements.txt
pandas==2.2.2
pyarrow==17.0.0
matplotlib==3.9.0
seaborn==0.13.2
missingno==0.5.2
scikit-learn==1.5.1
ipykernel==6.29.5
fastapi==0.112.0
uvicorn==0.30.5
lightgbm==4.5.0
python-dotenv==1.0.1
psycopg2==2.9.9
SQLAlchemy==2.0.32
pika==1.3.2
boto3==1.35.2
sphinx==7.2.6